streamlit_app/requirements.txt
requests~=2.32.3
streamlit~=1.41.1
numpy~=2.2.1
pandas~=2.2.3
pillow~=11.0.0
typing_extensions~=4.12.2